// src/app/modules/core/dmo/model/dmo.ts
export class MarketLoanRepaymentSchedule {
  financialYear: string;
  installmentDate: string;
  openingBalance: string;
  principal: string;
  interest: string;
  closingBalance: string;
}